[Test] replace ffmpeg to libav-tools and start refactoring test
In progress
