+++ .travis.yml
@@ -14,7 +14,6 @@ before_install:
 - sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu trusty main restricted universe multiverse"
 - sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu trusty-updates main restricted universe multiverse"
 - sudo add-apt-repository "deb http://archive.ubuntu.com/ubuntu trusty-backports main restricted universe multiverse"
-- sudo add-apt-repository ppa:mc3man/trusty-media -y
 - sudo apt-get update
 - sudo apt-get install $(cat install/files/deb-packages_requirements.txt)
 - sudo apt-get install libstdc++6
