+++ kalliope/players/pyaudioplayer/pyaudioplayer.py
@@ -4,7 +4,7 @@ import wave
 
 import pyaudio
 
-from kalliope.PlayerModule import PlayerModule
+from kalliope.core.PlayerModule import PlayerModule
 
 logging.basicConfig()
 logger = logging.getLogger("kalliope")
